Submissions/my_submission/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
openai>=1.0.0
python-dotenv>=1.0.0